refactor(analysis): isolate enclosure contracts
